+++ settings/index.js
@@ -1,15 +1,14 @@
 import _ from 'lodash';
 import React from 'react';
 import Settings from '@folio/stripes-components/lib/Settings';
-import PlaceholderSettings from './PlaceholderSettings';
-
+import MaterialTypesSettings from './MaterialTypesSettings';
 
 const pages = [
   {
-    route: 'placeholder',
-    label: 'Placeholder',
-    component: PlaceholderSettings,
-    perm: 'ui-inventory.settings.placeholder',
+    route: 'materialtypes',
+    label: 'Material Types',
+    component: MaterialTypesSettings,
+    perm: 'ui-inventory.settings.materialtypes',
   },
 ];