+++ CHANGELOG.md
@@ -47,6 +47,7 @@
 * Lint. Fixes UIIN-104.
 * Clone an item record. Fixes UIIN-27.
 * Clone a holding record. Fixes UIIN-26.
+* Add Material Types to Settings section. UIIN-87.
 
 ## [1.0.0](https://github.com/folio-org/ui-instances/tree/v1.0.0) (2017-09-08)
 [Full Changelog](https://github.com/folio-org/ui-items/compare/v0.0.1...v1.0.0)